refactor: many to one migration
