--- app/alembic/versions/0c7bd30b5a24_.py
@@ -1,29 +1,0 @@
-"""Add last logon.
-
-Revision ID: 0c7bd30b5a24
-Revises: f4a7fde509d4
-Create Date: 2024-02-22 08:06:09.716671
-
-Copyright (c) 2024 MultiFactor
-License: https://github.com/MultiDirectoryLab/MultiDirectory/blob/main/LICENSE
-"""
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '0c7bd30b5a24'
-down_revision = 'f4a7fde509d4'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('Users', sa.Column('lastLogon', sa.DateTime(timezone=True), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('Users', 'lastLogon')
-    # ### end Alembic commands ###
