--- app/alembic/versions/2c6556187ae3_path_in_directory.py
@@ -1,104 +1,0 @@
-"""Path in Directory
-
-Revision ID: 2c6556187ae3
-Revises: aab3b0e949d9
-Create Date: 2024-09-23 08:15:12.299848
-
-"""
-import sqlalchemy as sa
-from alembic import op
-from sqlalchemy import (
-    Column,
-    ForeignKey,
-    Integer,
-    String,
-    insert,
-    orm,
-    select,
-    update,
-)
-from sqlalchemy.dialects import postgresql
-from sqlalchemy.orm import declarative_base
-
-from models import Directory
-
-# revision identifiers, used by Alembic.
-revision = '2c6556187ae3'
-down_revision = 'aab3b0e949d9'
-branch_labels = None
-depends_on = None
-
-
-Base = declarative_base()
-
-
-class Path(Base):
-    """Directory path data."""
-
-    __tablename__ = "Paths"
-
-    id = Column(Integer, primary_key=True)  # noqa: A003
-    path = Column(postgresql.ARRAY(String), nullable=False, index=True)
-
-    endpoint_id = Column(Integer, ForeignKey('Directory.id'), nullable=False)
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('DirectoryPaths')
-    op.add_column('Directory', Column('path', postgresql.ARRAY(String())))
-    op.create_index(op.f('ix_Directory_path'), 'Directory', ['path'], unique=False)
-
-    bind = op.get_bind()
-    session = orm.Session(bind=bind)
-
-    subquery = (
-        select(Path.path)
-        .where(Path.endpoint_id == Directory.id)
-        .scalar_subquery())
-
-    session.execute(update(Directory).values(path=subquery))
-
-    op.alter_column('Directory', 'path', nullable=False)
-    op.drop_index('ix_Paths_path', table_name='Paths')
-    op.drop_index('lw_path', table_name='Paths')
-    op.drop_table('Paths')
-
-    op.execute(sa.text('CREATE INDEX lw_path ON "Directory" USING GIN(array_lowercase("path"));'))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('lw_path', table_name='Directory')
-    op.create_table('Paths',
-        sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-        sa.Column('path', postgresql.ARRAY(sa.VARCHAR()), autoincrement=False, nullable=False),
-        sa.Column('endpoint_id', sa.INTEGER(), autoincrement=False, nullable=False),
-        sa.ForeignKeyConstraint(['endpoint_id'], ['Directory.id'], name='Paths_endpoint_id_fkey'),
-        sa.PrimaryKeyConstraint('id', name='Paths_pkey')
-    )
-    op.execute(sa.text('CREATE INDEX lw_path ON "Paths" USING GIN(array_lowercase("path"));'))
-    op.create_table('DirectoryPaths',
-        sa.Column('dir_id', sa.INTEGER(), autoincrement=False, nullable=False),
-        sa.Column('path_id', sa.INTEGER(), autoincrement=False, nullable=False),
-        sa.ForeignKeyConstraint(['dir_id'], ['Directory.id'], name='DirectoryPaths_dir_id_fkey'),
-        sa.ForeignKeyConstraint(['path_id'], ['Paths.id'], name='DirectoryPaths_path_id_fkey'),
-        sa.PrimaryKeyConstraint('dir_id', 'path_id', name='DirectoryPaths_pkey')
-    )
-
-    bind = op.get_bind()
-    session = orm.Session(bind=bind)
-
-    subquery = select(
-        Directory.id.label('endpoint_id'),
-        Directory.path.label('path'))
-
-    insert_stmt = insert(Path).from_select(['endpoint_id', 'path'], subquery)
-    session.execute(insert_stmt)
-    session.commit()
-
-    op.create_index('ix_Paths_path', 'Paths', ['path'], unique=False)
-    op.drop_index(op.f('ix_Directory_path'), table_name='Directory')
-    op.drop_column('Directory', 'path')
-    # ### end Alembic commands ###
