--- app/alembic/versions/59e98bbd8ad8_change_gidnumber_for_users.py
@@ -1,42 +1,0 @@
-"""change gidNumber for users
-
-Revision ID: 59e98bbd8ad8
-Revises: 9c4f73699219
-Create Date: 2024-10-01 08:44:12.375329
-
-"""
-from alembic import op
-from sqlalchemy import select, update
-from sqlalchemy.orm import Session
-
-from models import Attribute, User
-
-# revision identifiers, used by Alembic.
-revision = '59e98bbd8ad8'
-down_revision = '9c4f73699219'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    bind = op.get_bind()
-    session = Session(bind=bind)
-
-    subquery = select(User.directory_id).as_scalar()
-    session.execute(
-        update(Attribute)
-        .values(value='513')
-        .where(
-            Attribute.directory_id.in_(subquery),
-            Attribute.name == 'gidNumber',
-        )
-        .execution_options(synchronize_session=False))
-    session.commit()
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
