--- app/alembic/versions/f4a7fde509d4_.py
@@ -1,158 +1,0 @@
-"""Init.
-
-Revision ID: f4a7fde509d4
-Revises:
-Create Date: 2023-11-13 16:07:06.467150
-
-Copyright (c) 2024 MultiFactor
-License: https://github.com/MultiDirectoryLab/MultiDirectory/blob/main/LICENSE
-"""
-import sqlalchemy as sa
-from alembic import op
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = 'f4a7fde509d4'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('Directory',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('parentId', sa.Integer(), nullable=True),
-    sa.Column('objectClass', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('whenCreated', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('whenChanged', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('depth', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['parentId'], ['Directory.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('parentId', 'name', name='name_parent_uc', postgresql_nulls_not_distinct=True)
-    )
-    op.create_index(op.f('ix_Directory_parentId'), 'Directory', ['parentId'], unique=False)
-    op.create_table('Policies',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('raw', postgresql.JSON(astext_type=sa.Text()), nullable=False),
-    sa.Column('netmasks', postgresql.ARRAY(postgresql.CIDR()), nullable=False),
-    sa.Column('enabled', sa.Boolean(), server_default=sa.text('true'), nullable=False),
-    sa.Column('priority', sa.Integer(), nullable=False),
-    sa.Column('mfa_status', sa.Enum('DISABLED', 'ENABLED', 'WHITELIST', name='mfaflags'), server_default='DISABLED', nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name'),
-    sa.UniqueConstraint('priority', deferrable='True', initially='DEFERRED', name='priority_uc')
-    )
-    op.create_index(op.f('ix_Policies_netmasks'), 'Policies', ['netmasks'], unique=True)
-    op.create_table('Settings',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('value', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_Settings_name'), 'Settings', ['name'], unique=False)
-    op.create_table('Attributes',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('value', sa.String(), nullable=False),
-    sa.Column('directoryId', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['directoryId'], ['Directory.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_Attributes_name'), 'Attributes', ['name'], unique=False)
-    op.create_table('Computers',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('directoryId', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['directoryId'], ['Directory.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('Groups',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('directoryId', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['directoryId'], ['Directory.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('Paths',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('path', postgresql.ARRAY(sa.String()), nullable=False),
-    sa.Column('endpoint_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['endpoint_id'], ['Directory.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_Paths_path'), 'Paths', ['path'], unique=False)
-    op.create_table('Users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('sAMAccountName', sa.String(), nullable=False),
-    sa.Column('userPrincipalName', sa.String(), nullable=False),
-    sa.Column('mail', sa.String(length=255), nullable=True),
-    sa.Column('displayName', sa.String(), nullable=True),
-    sa.Column('password', sa.String(), nullable=True),
-    sa.Column('directoryId', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['directoryId'], ['Directory.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('sAMAccountName'),
-    sa.UniqueConstraint('userPrincipalName')
-    )
-    op.create_table('DirectoryPaths',
-    sa.Column('dir_id', sa.Integer(), nullable=False),
-    sa.Column('path_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['dir_id'], ['Directory.id'], ),
-    sa.ForeignKeyConstraint(['path_id'], ['Paths.id'], ),
-    sa.PrimaryKeyConstraint('dir_id', 'path_id')
-    )
-    op.create_table('GroupMemberships',
-    sa.Column('group_id', sa.Integer(), nullable=False),
-    sa.Column('group_child_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['group_child_id'], ['Groups.id'], ),
-    sa.ForeignKeyConstraint(['group_id'], ['Groups.id'], ),
-    sa.PrimaryKeyConstraint('group_id', 'group_child_id')
-    )
-    op.create_table('PolicyMFAMemberships',
-    sa.Column('group_id', sa.Integer(), nullable=False),
-    sa.Column('policy_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['group_id'], ['Groups.id'], ),
-    sa.ForeignKeyConstraint(['policy_id'], ['Policies.id'], ),
-    sa.PrimaryKeyConstraint('group_id', 'policy_id')
-    )
-    op.create_table('PolicyMemberships',
-    sa.Column('group_id', sa.Integer(), nullable=False),
-    sa.Column('policy_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['group_id'], ['Groups.id'], ),
-    sa.ForeignKeyConstraint(['policy_id'], ['Policies.id'], ),
-    sa.PrimaryKeyConstraint('group_id', 'policy_id')
-    )
-    op.create_table('UserMemberships',
-    sa.Column('group_id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['group_id'], ['Groups.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['Users.id'], ),
-    sa.PrimaryKeyConstraint('group_id', 'user_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('UserMemberships')
-    op.drop_table('PolicyMemberships')
-    op.drop_table('PolicyMFAMemberships')
-    op.drop_table('GroupMemberships')
-    op.drop_table('DirectoryPaths')
-    op.drop_table('Users')
-    op.drop_index(op.f('ix_Paths_path'), table_name='Paths')
-    op.drop_table('Paths')
-    op.drop_table('Groups')
-    op.drop_table('Computers')
-    op.drop_index(op.f('ix_Attributes_name'), table_name='Attributes')
-    op.drop_table('Attributes')
-    op.drop_index(op.f('ix_Settings_name'), table_name='Settings')
-    op.drop_table('Settings')
-    op.drop_index(op.f('ix_Policies_netmasks'), table_name='Policies')
-    op.drop_table('Policies')
-    op.drop_index(op.f('ix_Directory_parentId'), table_name='Directory')
-    op.drop_constraint('name_parent_uc', 'Directory')
-    op.drop_table('Directory')
-    op.execute(sa.text('DROP TYPE mfaflags'))
-    # ### end Alembic commands ###
